Generalize liftPG to any monad
This is useful if you want to use connection from IO callback
Alternatively  we could export withConnection (former liftPG'). That
can actually be a good idea in addition to liftPG change.

Rename original liftPG to liftPG' as one specialized to IO for internal use.
And liftPG' renamed to withConnection, for better code readability.